split pending and unlocked withdrawals in separate components
